Extract OuroborosHub update transaction body into the marketplace layer
Behavior-neutral move of the gateway's _run_update closure and the
HubInstallResult serializer into ouroboros/marketplace/ouroboroshub.py
(run_hub_update / serialize_hub_install_result). The gateway keeps the
lifecycle-thread bridge and the shared review/deps orchestration as
injected callables so its monkeypatch seams stay authoritative. Frees
gateway/marketplace.py headroom for the adopt dispatch branch.
